Integrate OGPDHarvester plugin
From the inheritance point of view I thing it is the best to let the
OGPDHarvester derive from the GeminiCswHarvester class. I addition I
have adjusted the plugin path in the setup.py.

There was a typo in setting up the CSW client for the harvester which
made in invocation to the not present CSW server method.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>