Fix incompatibilities with different etree APIs
Due to varieties in different versions of the ElementTree APIs the
harvester tried to access the tag attribute in an ElementTree object.
The tag attribute, however, is only available in an Element object.

This leads to the problem, that the import stage will fail with every
fetched XML document with the error:

Error getting the CSW record with GUID ...

The actual exception message, which is swallowed, is: 'ElementTree'
object has no attribute 'tag'

Since there is already an exception handling which tries to deal with
these incompatibilies, I have switched the lines as Esra has explained
in the last meeting.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>