Add individual INSPIRE package writer and document
As a matter of fact the current Gemini XPath expressions differ from
what is known from the INSPIRE fields. This was noticed on the metadata
legal constraints, for instance the use limitation, a very general
constraint on the data usage, is used for determining the legal public
access in Gemini.

Hence, I have created a separate InspireDocument in the metadata model
and a separate write package method and import stage for the OGPD
Harvester.

For the legal constraints the fields accessConstraints, useConstraints
and otherConstraints are used now, as it is described in the German
translation for the ISO 19115 from GDI-DE as well.

Also I have changed the name for the OGPD Harvester from csw to ogpd in
order to differ from the GeminiCswHarvester

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>