Implement license mapping
For the OGD schema there are three fields to determine the terms of use
for a given resource. The 'license_id' is a certain enumeration type. The
'license_url' points, if given, to a description, respectively definition
of the license and in the field 'otherÄ additional information can given
specified.

The latter one is often required when a default mapping is not
applicable. So far this is often the case. Hence the default
'license_id' is 'other-closed' and additional information are harvested
from the INSPIRE constraint fields.

The license mapping is done in a separate Python module within the method
'translate_license_data'. If this method returns None it indicates to drop
the current CSW entry.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>