Start to map contact information
The necessary information for our contact information requirements are
already partly present, at least the name and the email. Hence I have
started to create the appropriate fields in the CKAN package.

The contacts field contains two objects with the keys role, name, url
email and address which are filled accordingly. Since we are aiming for
two contacts (publisher and owner/pointOfContact) these are initialized
and the rest of the data is added if present.

Especially the parsing of URL and address is missing.

Signed-off-by: Konrad Reiche <konrad.reiche@gmail.com>