Make licenses module conform to PEP8 conventions
PEP8 is a coding convention, especially for Python code comprising the
standard library in the main Python distribution. It is useful for
improving the consistency and readability.

http://www.python.org/dev/peps/pep-0008/

I have adapted that style for the licenses module and hopefully for the
following code as well. For similar reasons I have renamed the license
module.

Signed-off-by: Konrad Reiche <konrad.reiche@fokus.fraunhofer.de>