Implement Hamburg CKAN Harvester
There are two problems with harvesting CKAN data from Hamburg. Firstly,
the harvesting has to be done on API version 1, since the groups of the
datasets should be retrieved and currently the IDs are not the same
across different CKAN instances, hence the groups have to be retrieved
via their names.

Secondly, the groups of the Hamburg CKAN use hyphens (-) instead of
underscores (_). Until this is fixed, the group string is modified.

Signed-off-by: Konrad Reiche <konrad.reiche@fokus.fraunhofer.de>